[CHORE] moved files in the right order, made empty game  files
